+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [2.1.1](https://github.com/DenHeadless/DTModelStorage/releases/tag/2.1.1)
+
+#### Added
+* `insertSection(_:atIndex:)` that allows to insert `SectionModel` directly, with items and supplementary headers. 
+
 ## [2.1.0](https://github.com/DenHeadless/DTModelStorage/releases/tag/2.1.0)
 
 #### Updated
-* `StorageUpdate` class was rewritten from scratch using Swift `Set`.
-* `StorageUpdate` now contains `movedRowIndexPaths` and `movedSectionIndexes` properties.
+* `StorageUpdate` class was rewritten from scratch using Swift Set.
+* `StorageUpdate` now contains movedRowIndexPaths and movedSectionIndexes properties
+* All method names and properties, that contained `object` term in their name, have been renamed to read 'item' instead
 
 #### Fixed
 * `removeItems` method should no longer skip items, if their indexPath is reduced when removing previous item
@@ -16,7 +22,6 @@ All notable changes to this project will be documented in this file.
 * `moveCollectionViewSection:toSection` and `moveTableViewSection:toSection` have been replaced by `moveSection:toSection` method
 
 ## [2.0.0](https://github.com/DenHeadless/DTModelStorage/releases/tag/2.0.0)
-Released on 2015-09-13.
 
 Framework was completely rewritten from scratch in Swift 2.